Dart_B_Website/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic[email]==2.9.2
PyPDF2==3.0.1
httpx==0.27.2
python-dotenv==1.0.0

